aop实现自定义参数校验、整合EasyExcel实现文件的导入导出

+++ Minio_SpringBoot_Server/src/main/java/com/example/minio/entity/Account.java
@@ -1,9 +1,5 @@
 package com.example.minio.entity;
 
-import java.time.LocalDateTime;
-import java.io.Serializable;
-import java.util.Date;
-
 import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
@@ -12,6 +8,9 @@ import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.experimental.Accessors;
+
+import java.io.Serializable;
+import java.util.Date;
 
 /**
  * <p>
@@ -32,7 +31,7 @@ public class Account implements Serializable {
 
     @ApiModelProperty(value = "员工id")
     @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     @ApiModelProperty(value = "邮箱")
     private String email;
